+++ lib/ua_inspector/util/regex.ex
@@ -28,7 +28,7 @@ defmodule UAInspector.Util.Regex do
   end
 
   def build_engine_regex("Arachne"), do: build_engine_regex("Arachne\\/5\\.")
-  def build_engine_regex("Blink"), do: build_engine_regex("Chrome|Cronet")
+  def build_engine_regex("Blink"), do: build_engine_regex("Chr[o0]me|Cronet")
   def build_engine_regex("LibWeb"), do: build_engine_regex("LibWeb\\+LibJs")
 
   def build_engine_regex(name) do