+++ CHANGELOG.md
@@ -6,7 +6,7 @@
     - Added support to use "form factors" client hint header for improved detection
     - Default upstream database version is now `6.4.1`
     - Improved mobile app detection
-    - Upstream improvements for ARM platform detection have been integrated
+    - Upstream improvements for OS platform detection have been integrated
     - Upstream improvements for Blink engine version detection have been integrated
     - Upstream improvements for browser version detection have been integrated
     - Upstream improvements for Chromium version detection have been integrated