+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     - Upstream improvements for Windows version detection have been integrated
     - Upstream improvements for Linux client detection have been integrated
     - Upstream improvements for "Chrome OS" detection have been integrated
+    - Upstream improvements for "Fire OS" detection have been integrated
     - Upstream improvements for "LeafOS" detection have been integrated
     - Upstream improvements for "Puffin" browser detection have been integrated
     - Upstream improvements for "TV-Browser Internet" detection have been integrated